refactor: unify Docker-like label parsing in ZodTransforms
Move the canonical comma-aware label parser into `contracts/ZodTransforms`
as `parseLabelsString` and point `labelsStringSchema` at it. The client-side
`parseDockerLikeLabels` now delegates to that single implementation.

Previously `labelsStringSchema` used a naive `split(',')` reducer that
dropped continuation fragments, corrupting label values that contain commas
(e.g. `com.docker.compose.project.config_files` listing multiple files) --
the exact bug fixed for `parseDockerLikeLabels` in #381. Unifying on the
fixed parser removes that duplicated, divergent logic and gives every schema
consumer (nerdctl volume/inspect/event paths) the correct behavior.

Keeps the contracts -> clients layering intact: the client helper imports
from contracts, never the reverse.

Adds schema-level tests locking in comma- and equals-in-value handling.

Copilot-Session: ec267aca-911e-4a63-ac08-728d65ca1605
